+++ Assets/Modules/Upgrades/Code/Configs/LinearFormula.cs
@@ -5,7 +5,7 @@ namespace FPSShooter.Modules.Meta.Upgrades
     [Serializable]
     public sealed class LinearFormula : TableFormula
     {
-        public override int CalculateValue(int baseValue, int level)
+        public override float CalculateValue(int baseValue, int level)
         {
             return baseValue * level;
         }
