+++ Assets/Modules/Upgrades/Code/Configs/TableFormula.cs
@@ -2,6 +2,6 @@ namespace FPSShooter.Modules.Meta.Upgrades
 {
     public abstract class TableFormula
     {
-        public abstract int CalculateValue(int baseValue, int level);
+        public abstract float CalculateValue(int baseValue, int level);
     }
 }
