+++ Assets/Modules/Upgrades/Code/Configs/TableValue.cs
@@ -11,19 +11,19 @@ namespace FPSShooter.Modules.Meta.Upgrades
         [SerializeField] private int _baseValue = 1;
 
         [ReadOnly, SerializeField]
-        private int[] _values;
+        private float[] _values;
         public void FillLevelPrices(int maxLevel)
         {
             if (_formula == null) return;
             
-            _values = new int[maxLevel];
+            _values = new float[maxLevel];
             for (int i = 0; i < maxLevel; i++)
             {
                 _values[i] = _formula.CalculateValue(_baseValue, i + 1);
             }
         }
 
-        public int GetValue(int level)
+        public float GetValue(int level)
         {
             var index = level - 1;
             index = Mathf.Clamp(index, 0, _values.Length - 1);
